Write teacher and DBCperson super classes, tests pass for apprentice teacher class
