+++ senior_teacher.rb
@@ -1,4 +1,4 @@
-class SeniorTeacher
+ class SeniorTeacher
   attr_reader :age, :salary, :phase, :performance_rating, :target_raise
   attr_accessor :name
 
@@ -9,14 +9,14 @@ class SeniorTeacher
     @target_raise = 1000
   end
 
-  def offer_high_five
-    "High five!"
-  end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
+  # def offer_high_five
+  #   "High five!"
+  # end
+  #
+  # def set_phase(num)
+  #   @phase = num
+  #   "Cool, I've always wanted to teach phase #{num}!"
+  # end
 
   def teach_stuff
     response = ""
