+++ student.rb
@@ -1,3 +1,4 @@
+
 class Student
   attr_reader :age, :phase
   attr_accessor :name
@@ -8,9 +9,9 @@ class Student
     @name = options.fetch(:name, "")
   end
 
-  def offer_high_five
-    "High five!"
-  end
+  # def offer_high_five
+  #   "High five!"
+  # end
 
   def set_phase(num)
     response = ""
